// acheev-api/src/types/gqlTypings.generated.ts
export type Maybe<T> = T | undefined;
export type InputMaybe<T> = T | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  DateTime: any;
  JSON: any;
  Time: any;
};
// ...
export type Circuit = {
  __typename?: 'Circuit';
  archived: Scalars['Boolean'];
  completedAt?: Maybe<Scalars['DateTime']>;
  createdAt: Scalars['DateTime'];
  exercises: Array<Exercise>;
  id: Scalars['Int'];
  name: Scalars['String'];
  order: Scalars['Int'];
  programFacetId: Scalars['Int'];
  programId: Scalars['Int'];
  startedAt?: Maybe<Scalars['DateTime']>;
  workoutId: Scalars['Int'];
};
// ...
export enum CompletionModelType {
  Cicuit = 'CICUIT',
  Exercise = 'EXERCISE',
  ExerciseSet = 'EXERCISE_SET',
  Program = 'PROGRAM',
  ProgramFacet = 'PROGRAM_FACET',
  Workout = 'WORKOUT'
}
// ...
export type Exercise = {
  __typename?: 'Exercise';
  archived: Scalars['Boolean'];
  circuitId: Scalars['Int'];
  completedAt?: Maybe<Scalars['DateTime']>;
  createdAt: Scalars['DateTime'];
  description: Scalars['String'];
  exerciseSets: Array<ExerciseSet>;
  id: Scalars['Int'];
  name: Scalars['String'];
  order: Scalars['Int'];
  programFacetId: Scalars['Int'];
  programId: Scalars['Int'];
  restDurationSeconds: Scalars['Int'];
  startedAt?: Maybe<Scalars['DateTime']>;
  videoUrl: Scalars['String'];
  workoutId: Scalars['Int'];
};
// ...
export type ExerciseSet = {
  __typename?: 'ExerciseSet';
  archived: Scalars['Boolean'];
  circuitId: Scalars['Int'];
  completedAt?: Maybe<Scalars['DateTime']>;
  createdAt: Scalars['DateTime'];
  distance?: Maybe<Scalars['Int']>;
  durationSeconds: Scalars['Int'];
  exerciseId: Scalars['Int'];
  id: Scalars['Int'];
  maxPercAvailable?: Maybe<Scalars['Boolean']>;
  order: Scalars['Int'];
  percentageMax?: Maybe<Scalars['Float']>;
  programFacetId: Scalars['Int'];
  programId: Scalars['Int'];
  repCount?: Maybe<Scalars['Int']>;
  repsAvailable?: Maybe<Scalars['Boolean']>;
  startedAt?: Maybe<Scalars['DateTime']>;
  templateId?: Maybe<Scalars['Int']>;
  weight: Scalars['Float'];
  weightRelative: Scalars['Boolean'];
  weightUnit: WeightUnit;
  workoutId: Scalars['Int'];
};
// ...
export type ExerciseSetInput = {
  archived?: InputMaybe<Scalars['Boolean']>;
  circuitId?: InputMaybe<Scalars['Int']>;
  completedAt?: InputMaybe<Scalars['DateTime']>;
  distance?: InputMaybe<Scalars['Int']>;
  durationSeconds?: InputMaybe<Scalars['Int']>;
  exerciseId?: InputMaybe<Scalars['Int']>;
  maxPercAvailable?: InputMaybe<Scalars['Boolean']>;
  order?: InputMaybe<Scalars['Int']>;
  percentageMax?: InputMaybe<Scalars['Float']>;
  programFacetId?: InputMaybe<Scalars['Int']>;
  programId?: InputMaybe<Scalars['Int']>;
  repCount?: InputMaybe<Scalars['Int']>;
  repsAvailable?: InputMaybe<Scalars['Boolean']>;
  startedAt?: InputMaybe<Scalars['DateTime']>;
  weight?: InputMaybe<Scalars['Float']>;
  weightRelative?: InputMaybe<Scalars['Boolean']>;
  weightUnit?: InputMaybe<WeightUnit>;
  workoutId?: InputMaybe<Scalars['Int']>;
};
// ...
export enum FavoriteModelType {
  ProgramFacet = 'PROGRAM_FACET',
  Workout = 'WORKOUT'
}
// ...
export type Program = {
  __typename?: 'Program';
  archived: Scalars['Boolean'];
  createdAt: Scalars['DateTime'];
  id: Scalars['Int'];
  imageUrl: Scalars['String'];
  live: Scalars['Boolean'];
  name: Scalars['String'];
  programFacets: Array<ProgramFacet>;
};

export type ProgramFacet = {
  __typename?: 'ProgramFacet';
  archived: Scalars['Boolean'];
  createdAt: Scalars['DateTime'];
  description?: Maybe<Scalars['String']>;
  equipmentNeeded?: Maybe<Scalars['String']>;
  goals?: Maybe<Scalars['String']>;
  id: Scalars['Int'];
  imageUrl: Scalars['String'];
  isFavorited: Scalars['Boolean'];
  live: Scalars['Boolean'];
  name: Scalars['String'];
  order: Scalars['Int'];
  program: Program;
  programId: Scalars['Int'];
  videoUrl?: Maybe<Scalars['String']>;
};
// ...
export enum SkillLevel {
  Advanced = 'ADVANCED',
  Beginner = 'BEGINNER',
  Expert = 'EXPERT',
  Intermediate = 'INTERMEDIATE',
  Professional = 'PROFESSIONAL'
}
// ...
export enum WeightUnit {
  Kilos = 'KILOS',
  Pounds = 'POUNDS'
}

export type Workout = {
  __typename?: 'Workout';
  archived: Scalars['Boolean'];
  circuits: Array<Circuit>;
  completedAt?: Maybe<Scalars['DateTime']>;
  createdAt: Scalars['DateTime'];
  durationMinutes: Scalars['Int'];
  id: Scalars['Int'];
  imageUrl: Scalars['String'];
  isCompleted: Scalars['Boolean'];
  isFavorited: Scalars['Boolean'];
  name: Scalars['String'];
  order: Scalars['Int'];
  programFacet: ProgramFacet;
  programFacetId: Scalars['Int'];
  programId: Scalars['Int'];
  skillLevel: SkillLevel;
  startedAt?: Maybe<Scalars['DateTime']>;
  week: Scalars['Int'];
};